+++ src/main/java/manager/EnigmaManager.java
@@ -8,6 +8,9 @@ import repository.dao.EnigmaDAO;
 import repository.dao.EnigmaDAOSQL;
 import util.InputHandler;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
 public class EnigmaManager {
     private final EnigmaDAO dao = new EnigmaDAOSQL();
 
@@ -22,7 +25,7 @@ public class EnigmaManager {
         try {
             dao.insert(enigma);
             NotificationService.getInstance()
-                    .notifyObservers("Created new Enigma: " + enigma.getId());
+                    .notifyObservers("Created new Enigma: " + name);
         } catch (Exception e) {
             System.out.println("Error creating Enigma: " + e.getMessage());
         }
@@ -39,9 +42,16 @@ public class EnigmaManager {
     }
 
     public void updateEnigma() {
+        List<Enigma> enigmas = dao.findAll();
+        if (enigmas.isEmpty()) {
+            System.out.println("There are no Enigmas available to update.");
+            return;
+        }
+        System.out.println("Available Enigmas:");
+        enigmas.forEach(r -> System.out.println("  " + r.getId() + " → " + r.getName()));
         int id = InputHandler.readInt("Enter id: ");
         String name = InputHandler.readString("Enter new name");
-        int roomID = InputHandler.readInt("Enter new room ID");
+        int roomID = InputHandler.readOptionalInt("Enter new room ID (or press Enter to skip)");
         Theme theme = InputHandler.readEnum(Theme.class, "Enter new theme");
         String description = InputHandler.readString("Enter a new description:");
 
@@ -50,18 +60,29 @@ public class EnigmaManager {
         try {
             dao.update(enigma);
             NotificationService.getInstance()
-                    .notifyObservers("Updated Enigma: " + enigma.getId());
+                    .notifyObservers("Updated Enigma: " + name);
         } catch (Exception e) {
             System.out.println("Error updating Enigma: " + e.getMessage());
         }
     }
 
     public void deleteEnigma() {
+        List<Enigma> enigmas = dao.findAll();
+        if (enigmas.isEmpty()) {
+            System.out.println("There are no Enigmas available to delete.");
+            return;
+        }
+        System.out.println("Available Rooms:");
+        enigmas.forEach(r -> System.out.println("  " + r.getId() + " → " + r.getName()));
         int id = InputHandler.readInt("Enter Enigma ID to delete: ");
         try {
             dao.delete(id);
+            String name = enigmas.stream()
+                    .filter(r -> r.getId() == id)
+                    .map(Enigma::getName)
+                    .collect(Collectors.joining());
             NotificationService.getInstance()
-                    .notifyObservers("Deleted Enigma: " + id);
+                    .notifyObservers("Deleted Enigma: " + name);
         } catch (Exception e) {
             System.out.println("Error deleting Enigma: " + e.getMessage());
         }
